updates for linux-yocto 6.5
Upstream Linux kernel 6.4 went EOL sooner than expected, so upstream oe-core
introduced a 6.5 recipe for linux-yocto into master. Now oe-core master
supports 6.1, 6.4, and 6.5.

Starting with 6.5 the location of the 32-bit Arm dts files has moved into
vendor/soc-specific directories (bringing it in-line with the organization of
the 64-bit Arm dts files). While this cleanup is nice, it causes havoc with
the yocto. If the user wants to use a pre-6.5 kernel then the
KERNEL_DEVICETREE would be am335x-pocketbeagle.dtb, but if the user wants to
use a 6.5 and later kernel then the KERNEL_DEVICETREE would be
ti/omap/am335x-pocketbeage.dtb. The KERNEL_DEVICETREE is specified in
conf/machine.conf files, which don't know or care which version of the kernel
the user wants to use.

meta-ti has a convoluted, but working, solution. I'm not as interested
in bringing in such a heavy solution, especially considering this extra
plumbing is only required in branches where pre-6.5 and 6.5-and-later
versions are mixed together. So for the time-being I'll fixup the build to
expect users to use the 6.5 kernel.

Also, some of the labels in the dts file itself have changed, so this patch
tweaks the in-layer kernel patches to match.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>